use clustering algo n_jobs handling as a mixin pattern
