+++ pyemma/coordinates/clustering/interface.py
@@ -28,6 +28,7 @@ import os
 
 import numpy as np
 from pyemma._base.model import Model
+from pyemma._base.parallel import NJobsMixIn
 from pyemma._ext.sklearn.base import ClusterMixin
 from pyemma.coordinates.clustering import regspatial
 from pyemma.coordinates.data._base.transformer import StreamingEstimationTransformer
@@ -39,7 +40,7 @@ from six.moves import range, zip
 
 @fix_docs
 @aliased
-class AbstractClustering(StreamingEstimationTransformer, Model, ClusterMixin):
+class AbstractClustering(StreamingEstimationTransformer, Model, ClusterMixin, NJobsMixIn):
 
     """
     provides a common interface for cluster algorithms.
@@ -63,57 +64,6 @@ class AbstractClustering(StreamingEstimationTransformer, Model, ClusterMixin):
         self._overwrite_dtrajs = False
         self._index_states = []
         self.n_jobs = n_jobs
-
-    @property
-    def n_jobs(self):
-        """ Returns number of jobs/threads to use during assignment of data.
-
-        Returns
-        -------
-        If None it will return number of processors /or cores or the setting of 'OMP_NUM_THREADS' env variable.
-
-        Notes
-        -----
-        By setting the environment variable 'OMP_NUM_THREADS' to an integer,
-        one will override the default argument of n_jobs (currently None).
-        """
-        assert isinstance(self._n_jobs, int)
-        return self._n_jobs
-
-    @n_jobs.setter
-    def n_jobs(self, val):
-        """ set number of jobs/threads to use via assignment of data.
-        Parameters
-        ----------
-        val: int or None
-            a positive int for the number of jobs. Or None to usage all available resources.
-
-        Notes
-        -----
-
-        """
-        from pyemma.util.reflection import get_default_args
-        def_args = get_default_args(self.__init__)
-
-        # default value from constructor?
-        if val == def_args['n_jobs']:
-            omp_threads_from_env = os.getenv('OMP_NUM_THREADS', None)
-            import psutil
-            n_cpus = psutil.cpu_count()
-            if omp_threads_from_env:
-                try:
-                    self._n_jobs = int(omp_threads_from_env)
-                    self.logger.info("number of threads obtained from env variable"
-                                     " 'OMP_NUM_THREADS'=%s" % omp_threads_from_env)
-                except ValueError as ve:
-                    self.logger.warning("could not parse env variable 'OMP_NUM_THREADS'."
-                                        " Value='{}'. Error={}. Will use {} jobs."
-                                        .format(omp_threads_from_env, ve, n_cpus))
-                    self._n_jobs = n_cpus
-            else:
-                self._n_jobs = n_cpus
-        else:
-            self._n_jobs = int(val)
 
     @property
     @alias('labels_')  # sk-learn compat.
